fix a problem that syscall args are incorrectly serialized (as discussed, we need to make this arch specific); fix sec_hw mailbox limit counter are u16 (again, need to make it arch specific); debugging prints

+++ arch/sec_hw/include/arch/mailbox_runtime.h
@@ -8,7 +8,7 @@ void close_runtime(void);
 void runtime_core(void);
 
 void mailbox_change_queue_access(uint8_t queue_id, uint8_t access, uint8_t proc_id);
-int mailbox_attest_queue_access(uint8_t queue_id, uint8_t access, uint8_t count);
+int mailbox_attest_queue_access(uint8_t queue_id, uint8_t access, uint16_t count);
 void runtime_recv_msg_from_queue(uint8_t *buf, uint8_t queue_id);
 void runtime_send_msg_on_queue(uint8_t *buf, uint8_t queue_id);
 void runtime_recv_msg_from_queue_large(uint8_t *buf, uint8_t queue_id);
